requirements.txt
streamlit
transformers
onnxruntime
tensorflow
numpy
Pillow